refactor: 🧩 Move queries and modes to collections
