+++ src/collections/collections.js
@@ -3,7 +3,7 @@ export * as asserts from "./asserts.js";
 export * as fetcherTemplates from "./fetcherTemplates.js";
 export * as fetchers from "./fetchers/fetchers.js";
 export * as fetcherSenders from "./fetcherSenders.js";
-export * as modes from "../utils/modes.js";
-export * as queries from "../utils/querys.js";
+export * as modes from "./modes.js";
+export * as queries from "./querys.js";
 export * as signals from "./signals.js";
 export * as requests from "./requests/requests.js";
