+++ src/utils/api.js
@@ -1,6 +1,6 @@
 import { asserts } from "../collections/collections.js";
 import { batch, createEffect, createSignal, onCleanup, untrack } from "solid-js";
-import * as queries from "./querys";
+import * as queries from "../collections/querys";
 import { getDates } from "./dates";
 import { TokenBucket } from "./TokenBucket";
 const DEBUG = location.origin.includes("localhost");
